Move undistract app to /app
