+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.Undistract"
-        android:name=".UndistractApp"
+        android:name=".app.UndistractApp"
         tools:targetApi="31">
         <activity
             android:name=".MainActivity"
