+++ app/src/main/java/com/undistract/BlockerViewModel.kt
@@ -3,7 +3,7 @@ package com.undistract
 import android.app.Application
 import android.content.Context
 import androidx.lifecycle.AndroidViewModel
-import com.undistract.UndistractApp
+import com.undistract.app.UndistractApp
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.StateFlow
